+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @webpack-blocks/core - Changelog
 
+## master
+
+- Change `match()` to passes all [webpack rule options](https://webpack.js.org/configuration/module/) to context. ([#250](https://github.com/andywer/webpack-blocks/pull/250))
+
 ## 1.0.0-beta.2
 
 - More useful error message when passing invalid blocks to `createConfig()` ([#171](https://github.com/andywer/webpack-blocks/issues/171))